added dashboard, search, firebase functions, image, responsiveness, functioning as counsellor and as student, college detail view page
