+++ package.json
@@ -18,6 +18,7 @@
     "react-icons": "^5.0.1",
     "react-loader-spinner": "^6.1.6",
     "react-router-dom": "^6.22.3",
+    "react-scroll": "^1.9.0",
     "react-scroll-to-top": "^3.0.0",
     "react-spinners": "^0.13.8",
     "typewriter-effect": "^2.21.0",
