+++ src/App.jsx
@@ -2,7 +2,6 @@ import { Outlet } from "react-router-dom";
 import "./styles/App.css";
 
 const App = () => {
-
   return (
     <div className="App">
       <Outlet />
@@ -10,9 +9,4 @@ const App = () => {
   );
 };
 
-
-
 export default App;
-
-
-
