--- src/components/Dashboard/Dashboard.css
@@ -1,420 +1,0 @@
-.App {
-  text-align: center;
-  background-color: #CAE7FE;
-}
-
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 10px;
-  color: white;
-  background-color: #CAE7FE;
-}
-
-.logo img {
-  width: 75px;
-  padding-left: 15px;
-}
-
-.menu {
-  flex-direction: column;
-  align-items: center;
-  top: 70px;
-  width: 100%;
-}
-
-.menu.show {
-  display: flex;
-}
-
-.menu ul {
-  list-style: none;
-  padding: 0;
-  display: flex;
-  gap: 30px;
-  float: right;
-  padding: 0px 30px;
-  align-items: center;
-  background-color: #f4f6fcb8;
-  border-top-left-radius: 50px;
-  border-bottom-left-radius: 50px;
-}
-
-.menu li {
-  margin: 15px 0;
-}
-
-.profile_btn {
-  background-color: var(--btnColor);
-  border-radius: 15px;
-  color: #F4F6FC;
-  text-align: center;
-  padding: 10px 50px;
-  margin-right: 0.5vw;
-  border: none;
-  font-size: 25px;
-  float: right;
-  cursor: pointer;
-}
-
-.menu a {
-  text-decoration: none;
-  color: #12229d;
-  font-size: 22px;
-}
-
-.hamburger {
-  cursor: pointer;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  width: 30px;
-  height: 25px;
-  display: none;
-}
-
-.bar {
-  width: 30px;
-  height: 3px;
-  background-color: white;
-  transition: 0.4s;
-}
-
-.bar.open {
-  transform: rotate(-45deg) translate(-4px, 6px);
-}
-
-.bar.open:nth-child(2) {
-  opacity: 0;
-}
-
-.bar.open:nth-child(3) {
-  transform: rotate(45deg) translate(-4px, -6px);
-}
-
-@media (max-width: 960px) {
-  .menu {
-    display: none;
-  }
-
-  .menu.show {
-    display: block;
-  }
-
-  .navbar {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 10px;
-    color: #12229d;
-  }
-
-  .menu {
-    display: none;
-    flex-direction: column;
-    align-items: center;
-    position: absolute;
-    top: 260px;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    width: 95%;
-    background-color: #CAE7FE;
-  }
-
-  .menu.show {
-    display: flex;
-    background: #cae7fe45;
-    border-radius: 10px;
-    padding: 75px;
-    box-shadow: 0 0 10px #cae7fe;
-    backdrop-filter: blur(10px);
-  }
-
-  .menu ul {
-    flex-direction: column;
-    list-style: none;
-    padding: 0;
-    background-color: transparent;
-    border: none;
-    border-radius: 0;
-  }
-
-  .menu li {
-    margin: 15px 0;
-  }
-
-  .menu a {
-    text-decoration: none;
-    color: #12229d;
-    font-size: 18px;
-  }
-
-  .hamburger {
-    display: flex;
-    margin-right: 30px;
-  }
-
-  .bar {
-    background-color: #12229d;
-  }
-
-  .dot {
-    display: none;
-  }
-}
-
-.maintxt {
-  background-color: #CAE7FE;
-}
-
-.maintxt h1 {
-  width: 66vw;
-  font-family: 'ABeeZee';
-  font-size: 70px;
-  color: #050A30;
-  margin: 0 auto;
-  padding-top: 64px;
-}
-
-.blue {
-  color: #3C38B9;
-}
-
-.maintxt p {
-  font-family: 'ABeeZee';
-  font-size: 22px;
-  color: #050A30;
-  text-align: center;
-  margin: 30px 0 0 0;
-  padding-bottom: 30px;
-}
-
-.search {
-  width: 70vw;
-  background-color: #5CB6F9;
-  padding: 10px;
-  border-radius: 15px;
-  margin: 0 auto;
-  margin-top: 50px;
-  margin-bottom: 50px;
-  display: flex;
-}
-
-.vl {
-  position: relative;
-  top: 1px;
-  right: -25px;
-  border-left: 2px solid #306094;
-  height: 70px;
-}
-
-.search img {
-  width: 30px;
-  height: 36px;
-  position: relative;
-  top: 18px;
-  right: -10px;
-  color: #306094;
-}
-
-.search input {
-  width: 78%;
-  height: 70px;
-  background-color: #5CB6F9;
-  appearance: none;
-  border: none;
-  font-size: 25px;
-  font-family: 'ABeeZee';
-  color: #F4F6FC;
-  padding-left: 35px;
-  margin-left: 40px;
-  margin-right: 15px;
-}
-
-.search input::placeholder {
-  color: #306094;
-}
-
-.search button {
-  background-color: var(--btnColor);
-  border-radius: 15px;
-  color: #F4F6FC;
-  text-align: center;
-  padding: 20px 50px 20px 50px;
-  margin-right: 0.5vw;
-  border: none;
-  font-size: 25px;
-  float: right;
-  cursor: pointer;
-}
-
-.search button:hover {
-  background-color: #000000
-}
-
-.s_bar_c {
-  display: contents;
-}
-
-@media (max-width: 1050px) {
-  .s_bar_c {
-    display: flex;
-    margin-bottom: 15px;
-  }
-
-  .search {
-    flex-direction: column;
-    background-color: #ffffff;
-  }
-
-  .search img {
-    display: none;
-  }
-
-  .vl {
-    display: none;
-  }
-
-  .search button {
-    width: fit-content;
-    margin: 0 auto;
-  }
-
-  .search input {
-    margin: 0;
-    padding: 0;
-    width: -webkit-fill-available;
-    border-radius: 15px;
-    padding-left: 12px;
-  }
-}
-
-@media (max-width: 960px) {
-  .maintxt h1 {
-    font-size: 55px;
-  }
-
-  .search input {
-    height: 60px;
-    font-size: 22px;
-    margin: 0;
-    padding: 0;
-    width: -webkit-fill-available;
-    border-radius: 15px;
-    padding-left: 12px;
-  }
-
-  .search button {
-    height: 60px;
-    padding: 0 30px;
-  }
-}
-
-.navigator {
-  font-size: 40px;
-  font-family: 'ABeeZee';
-  font-weight: 600;
-  width: 80vw;
-  margin: auto;
-  background-color: #CAE8FF;
-}
-
-.nearby {
-  float: left;
-  color: #050A30;
-}
-
-.seeall {
-  float: right;
-  color: #303B5E;
-}
-
-.colleges {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  margin: auto;
-  width: 80vw;
-  margin-top: 120px;
-  padding-left: 50px;
-  padding-right: 50px;
-  justify-content: space-between;
-  row-gap: 30px;
-  margin-bottom: 50px;
-}
-
-.college {
-  font-family: 'ABeeZee';
-  width: 300px;
-  height: 170px;
-  border-radius: 12px;
-  background-color: #5CB6F9;
-  padding: 30px;
-  margin: auto;
-}
-
-.up {
-  display: flex;
-}
-
-.up img {
-  width: 130px;
-  height: 130px;
-  border-radius: 12px;
-  margin-right: 20px;
-}
-
-.context p {
-  font-size: 20px;
-  font-family: 'ABeeZee';
-  margin-bottom: 20px;
-}
-
-.context span {
-  font-size: 14px;
-  background-color: #F4F6FC;
-  padding-top: 10px;
-  padding-bottom: 10px;
-  padding-left: 30px;
-  padding-right: 30px;
-  border-radius: 20px;
-  color: #3E3ABA;
-  font-family: 'ABeeZee';
-}
-
-.down {
-  display: flex;
-  margin-top: 20px;
-  justify-content: space-between;
-}
-
-.ctc {
-  font-family: 'ABeeZee';
-  font-size: 18px;
-  color: #050A30;
-}
-
-.time {
-  font-size: 15px;
-  font-family: 'ABeeZee';
-  color: #050a3079;
-}
-
-@media (max-width: 1400px) {
-  .colleges {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    row-gap: 30px;
-  }
-
-  .college {
-    margin: auto;
-  }
-}
-
-@media (max-width: 960px) {
-  .colleges {
-    grid-template-columns: 1fr;
-  }
-}
