--- src/components/Dashboard/colleges.json
@@ -1,86 +1,0 @@
-[
-    {
-        "name": "Ymca Jc Bose",
-        "location": "Faridabad",
-        "imageURL": "https://i.postimg.cc/cg4NH0pV/image.png",
-        "ctc": "Hightest CTC - 29LPA",
-        "time": "3 min ago"
-    },
-    {
-        "name": "Manav Rachna",
-        "location": "Faridabad",
-        "imageURL": "https://i.postimg.cc/cg4NH0pV/image.png",
-        "ctc": "Hightest CTC - 23LPA",
-        "time": "3 min ago"
-    },
-    {
-        "name": "EIT",
-        "location": "Faridabad",
-        "imageURL": "https://i.postimg.cc/cg4NH0pV/image.png",
-        "ctc": "Hightest CTC - 17LPA",
-        "time": "3 min ago"
-    },
-    {
-        "name": "Ymca Jc Bose",
-        "location": "Faridabad",
-        "imageURL": "https://i.postimg.cc/cg4NH0pV/image.png",
-        "ctc": "Hightest CTC - 29LPA",
-        "time": "3 min ago"
-    },
-    {
-        "name": "Manav Rachna",
-        "location": "Faridabad",
-        "imageURL": "https://i.postimg.cc/cg4NH0pV/image.png",
-        "ctc": "Hightest CTC - 23LPA",
-        "time": "3 min ago"
-    },
-    {
-        "name": "EIT",
-        "location": "Faridabad",
-        "imageURL": "https://i.postimg.cc/cg4NH0pV/image.png",
-        "ctc": "Hightest CTC - 17LPA",
-        "time": "3 min ago"
-    },
-    {
-        "name": "Ymca Jc Bose",
-        "location": "Faridabad",
-        "imageURL": "https://i.postimg.cc/cg4NH0pV/image.png",
-        "ctc": "Hightest CTC - 29LPA",
-        "time": "3 min ago"
-    },
-    {
-        "name": "Manav Rachna",
-        "location": "Faridabad",
-        "imageURL": "https://i.postimg.cc/cg4NH0pV/image.png",
-        "ctc": "Hightest CTC - 23LPA",
-        "time": "3 min ago"
-    },
-    {
-        "name": "EIT",
-        "location": "Faridabad",
-        "imageURL": "https://i.postimg.cc/cg4NH0pV/image.png",
-        "ctc": "Hightest CTC - 17LPA",
-        "time": "3 min ago"
-    },
-    {
-        "name": "Ymca Jc Bose",
-        "location": "Faridabad",
-        "imageURL": "https://i.postimg.cc/cg4NH0pV/image.png",
-        "ctc": "Hightest CTC - 29LPA",
-        "time": "3 min ago"
-    },
-    {
-        "name": "Manav Rachna",
-        "location": "Faridabad",
-        "imageURL": "https://i.postimg.cc/cg4NH0pV/image.png",
-        "ctc": "Hightest CTC - 23LPA",
-        "time": "3 min ago"
-    },
-    {
-        "name": "EIT",
-        "location": "Faridabad",
-        "imageURL": "https://i.postimg.cc/cg4NH0pV/image.png",
-        "ctc": "Hightest CTC - 17LPA",
-        "time": "3 min ago"
-    }
-]
