+++ src/index.css
@@ -1,26 +1,24 @@
-:root{
-    --pageColor: #ffffff;
-    --secPageColor: #6cbdfa;
-    --white: #ffffff;
-    --btnColor: #12229d;
-    font-size: 62.5%;
-    overflow-x: hidden;
+:root {
+  --pageColor: #ffffff;
+  --secPageColor: #6cbdfa;
+  --white: #ffffff;
+  --btnColor: #12229d;
+  font-size: 62.5%;
+  overflow-x: hidden;
 }
 
-body, html{
-    margin: 0;
-    min-width: 100vw;
+body,
+html {
+  margin: 0;
+  min-width: 100vw;
 }
 
-#root, .App{
-    background-color: var(--pageColor);
-    overflow-x: hidden;
+#root,
+.App {
+  background-color: var(--pageColor);
+  overflow-x: hidden;
 
-    height: 100%;
-    
-    width: 100%;
+  height: 100%;
+
+  width: 100%;
 }
-
-
-
-
